Make applyRefactorings take [String] instead of ([Extension], [Extension])
