+++ apply-refact.cabal
@@ -44,9 +44,6 @@ library
                , filemanip >= 0.3.6.3 && < 0.4
                , unix-compat >= 0.5.2
                , directory >= 1.3
-  if impl(ghc >= 8.8)
-    build-depends:
-      ghc-lib-parser-ex >= 8.10.0.16
   hs-source-dirs:      src
   default-language:    Haskell2010
 
@@ -80,9 +77,6 @@ executable refactor
                , unix-compat
                , filepath
                , transformers
-  if impl(ghc >= 8.8)
-    build-depends:
-      ghc-lib-parser-ex >= 8.10.0.16
 
 Test-Suite test
   type:                exitcode-stdio-1.0
@@ -117,6 +111,3 @@ Test-Suite test
                , filepath
                , silently
                , transformers
-  if impl(ghc >= 8.8)
-    build-depends:
-      ghc-lib-parser-ex >= 8.10.0.16
