--- cabal.project
@@ -1,2 +1,0 @@
-packages: *.cabal
-constraints: ghc-lib-parser-ex -auto +no-ghc-lib
